Fixed #1490 and #1520
